requirements.txt
requests
pytesseract
pillow
opencv-python-headless
discord-webhook
numpy